% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dme_genes-data.R
\docType{data}
\name{Dme_genes}
\alias{Dme_genes}
\title{Dme_genes}
\format{
An object of class \code{data.frame} with 34920 rows and 4 columns.
}
\source{
\href{https://www.ensembl.org/index.html}{ensembl biomaRt}
}
\usage{
Dme_genes
}
\description{
A gene ID table for \emph{D. melanogaster}.
}
\details{
This table was built directly from the ensembl biomaRt.
}
\section{Format}{

A dataframe with the folowing columns :
\itemize{
\item \code{fb_id}: FbGene ID (\emph{e.g} \code{FBgn0038257})
\item \code{transcript_id}: Transcript ID (\emph{e.g} \code{FBtr0336877})
\item \code{public_name}: Gene name (\emph{e.g} \code{smp-30})
\item \code{transcript_length}: Transcript length from 5'UTR start to 3'UTR end (\emph{e.g} \code{1102})
}
}

\keyword{datasets}
