% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_refs.R
\name{.plot_refs}
\alias{.plot_refs}
\title{Plot Reference Timelines}
\usage{
.plot_refs()
}
\description{
Plots the coverage of the \code{drosoRef} references (requires ggplot2)
}
\examples{
\donttest{
.plot_refs()
}

}
