% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_list-data.R
\docType{data}
\name{ref_list}
\alias{ref_list}
\title{List of reference in the wormRef data-package}
\format{
An object of class \code{data.frame} with 1 rows and 3 columns.
}
\usage{
ref_list
}
\description{
This object lists the pre-built references available for use with RAPToR.
}
\section{Format}{


A dataframe with the name of the reference, the organism, and a brief description.
More information on the references is given on their documentation pages
(\emph{e.g} \code{?Dme_embryo}).
}

\keyword{datasets}
